Fix a bug dumping colors not set in the scheme
CSApproxSnapshot should make a snapshot of the currently used colors,
rather than the colors defined in the colorscheme.  But, when
CSApproxSnapshot() changed &t_Co, the scheme was reloaded, causing :hi
clear to run and delete all colors.  Definitely not desired.  Stop that
from happening in the same way as it is prevented in CSApprox().