Work around the vim bug introduced in 8.0.0201
If a highlight group was cleared, versions prior to 8.0.0201 would
return -1 or "" for all attributes (except "name") for that group.
8.0.0201 would incorrectly return attributes for an entirely unrelated
group instead (the next non-cleared one in ID order).

We can work around this for builtin highlight group in particular,
because we know what their names should be and we can detect when they
have been cleared.